requirements.txt
discord.py==2.3.2
websocket-client==1.7.0
python-dotenv==1.0.0
requests==2.31.0 